+++ src/App.jsx
@@ -33,9 +33,9 @@ function App() {
           <Route path={PATH.MY_CALENDAR} element={<PageRoutes.MyCalendarPage />} />
           <Route path={PATH.MY_EMAIL_INQUIRE} element={<PageRoutes.MyEmailInquirePage />} />
           <Route path={PATH.MY_1V1_INQUIRE} element={<PageRoutes.My1v1InquirePage />} />
-          <Route path={PATH.INSTALLMENT_LIST} element={<PageRoutes.ProductInsListPage />} />
           {/* <Route path={PATH.DEPOSIT_LIST} element={<PageRoutes.ProductDepListPage />} /> */}
-          <Route path={PATH.DEPOSIT_LIST} element={<PageRoutes.DepositListPage />} />
+          <Route path={PATH.DEPOSIT_LIST} element={<PageRoutes.ListDepositPage />} />
+          <Route path={PATH.INSTALLMENT_LIST} element={<PageRoutes.ListInstallmentPage />} />
           <Route path={`${PATH.PRODUCT_COMPARE}/:ctg`} element={<PageRoutes.ProductComparePage />} />
           <Route path={`${PATH.PRODUCT_DETAIL}/:prdId`} element={<PageRoutes.ProductDetailPage />} />
           <Route path={PATH.PRODUCT_DETAIL} element={<Navigate to={`${PATH.PRODUCT_DETAIL}/1`} replace />} />
