--- src/pages/DepositListPage/index.js
@@ -1,3 +1,0 @@
-import DepositListPage from './DepositListPage';
-
-export default DepositListPage;
