+++ src/pages/ListDepositPage/ListDepositPage.module.scss
@@ -1,4 +1,4 @@
-/* src/pages/ProductInsListPage/ProductInsListPage.scss */
+/* src/pages/ListDepositPage/ListDepositPage.scss */
 
 .listSection {
   font-size: 16px;
@@ -173,7 +173,6 @@
             width: 48%;
             border: 1px solid var(--divider-border);
             flex: 1;
-            // background-color: var(--white);
             cursor: pointer;
             text-align: center;
             height: 40px;
@@ -311,6 +310,7 @@
 
         .productCompareBtn {
           width: 60px;
+          font-size: 15px;
           aspect-ratio: 1 / 1;
           background: var(--bg2);
           cursor: pointer;
