+++ src/pages/ListDepositPage/index.js
@@ -1,0 +1,3 @@
+import ListDepositPage from './ListDepositPage';
+
+export default ListDepositPage;
