+++ src/pages/ListInstallmentPage/index.js
@@ -1,0 +1,3 @@
+import ListInstallmentPage from './ListInstallmentPage';
+
+export default ListInstallmentPage;
