--- src/pages/ProductInsListPage/index.js
@@ -1,3 +1,0 @@
-import ProductInsListPage from './ProductInsListPage';
-
-export default ProductInsListPage;
